Add unittests for copy, move
