+++ packages/web-app-files/src/helpers/resource/index.ts
@@ -4,4 +4,4 @@ export * from './privatePreviewBlob'
 export * from './publicPreviewUrl'
 export * from './resource'
 export * from './sameResource'
-export * from './move'
+export * from './copyMove'
